+++ packages/cli/src/services/BuiltinCommandLoader.ts
@@ -28,7 +28,6 @@ import { restoreCommand } from '../ui/commands/restoreCommand.js';
 import { statsCommand } from '../ui/commands/statsCommand.js';
 import { themeCommand } from '../ui/commands/themeCommand.js';
 import { toolsCommand } from '../ui/commands/toolsCommand.js';
-import { createPromptCommands } from '../ui/commands/promptCommands.js';
 
 /**
  * Loads the core, hard-coded slash commands that are an integral part
@@ -67,7 +66,6 @@ export class BuiltinCommandLoader implements ICommandLoader {
       statsCommand,
       themeCommand,
       toolsCommand,
-      ...createPromptCommands(this.config),
     ];
 
     return allDefinitions.filter((cmd): cmd is SlashCommand => cmd !== null);
