+++ packages/core/src/index.ts
@@ -46,6 +46,9 @@ export * from './services/ideContext.js';
 export * from './tools/tools.js';
 export * from './tools/tool-registry.js';
 
+// Export prompt logic
+export * from './prompts/mcp-prompts.js';
+
 // Export specific tool logic
 export * from './tools/read-file.js';
 export * from './tools/ls.js';
